challenges/regionals-2024-25/web/BallotScanner/requirements.txt
Flask
Flask-Session
flask_sqlalchemy
pillow
pyzbar
boxdetect
cv2
numpy